requirements.txt
llama-cpp-python==0.2.56
numpy>=1.24.3
sounddevice>=0.4.6
soundfile>=0.12.1
PyAudio>=0.2.13
openai-whisper>=20231117
librosa>=0.10.1
kokoro-onnx>=1.0.0
PyYAML>=6.0.1 